agents/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
groq
